requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
psycopg2-binary
google-generativeai
Pillow
requests
